make delete confirmation actually work (broke it with adding extra handler that could run before the confirm handler)

+++ app/admin/static/js/events.js
@@ -7,8 +7,8 @@ $(document).ready(function() {
         $("#delete_images-input").val("");
     });
 
-    $("#delete_post-input").on("click", async function() {
+    $("#delete_post-input").on("click", confirmBtn(async function(e) {
         const respJson = await fetchWrapper(window.location.href, {method: "DELETE"});
         doAjaxResponseForm(respJson, e);
     });
-});
+}));
